requirements.txt
perceval-quandela==0.12.0
torch==2.*
torchvision~=0.19
tqdm==4.67.0
jupyter
pandas==2.2.3
